# backend/openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/warehouse/materials/balance": {
      "get": {
        "operationId": "WarehouseController_getMaterialBalance",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get material stock balances",
        "tags": [
          "Warehouse"
        ]
      }
    },
    "/api/v1/warehouse/materials/receipt": {
      "post": {
        "operationId": "WarehouseController_postMaterialReceipt",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Приход материала (закуп): остаток + цена + каскад в себестоимость",
        "tags": [
          "Warehouse"
        ]
      }
    },
    "/api/v1/warehouse/materials/{id}/movements": {
      "get": {
        "operationId": "WarehouseController_getMaterialMovements",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "История движений материала — за сколько и когда покупали",
        "tags": [
          "Warehouse"
        ]
      }
    },
    "/api/v1/warehouse/receipts": {
      "get": {
        "operationId": "WarehouseController_getReceipts",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Журнал приходов материалов",
        "tags": [
          "Warehouse"
        ]
      }
    },
    "/api/v1/warehouse/materials/movements": {
      "post": {
        "operationId": "WarehouseController_postMaterialMovement",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Record material stock movement",
        "tags": [
          "Warehouse"
        ]
      }
    },
    "/api/v1/warehouse/finished-goods/balance": {
      "get": {
        "operationId": "WarehouseController_getFGBalance",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get finished goods stock balance",
        "tags": [
          "Warehouse"
        ]
      }
    },
    "/api/v1/warehouse/finished-goods/movements": {
      "post": {
        "operationId": "WarehouseController_postFGMovement",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Record finished goods movement",
        "tags": [
          "Warehouse"
        ]
      }
    },
    "/api/v1/purchase-requests": {
      "get": {
        "operationId": "PurchaseRequestsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List purchase requests",
        "tags": [
          "Purchase Requests"
        ]
      },
      "post": {
        "operationId": "PurchaseRequestsController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create purchase request",
        "tags": [
          "Purchase Requests"
        ]
      }
    },
    "/api/v1/purchase-requests/{id}/approve": {
      "post": {
        "operationId": "PurchaseRequestsController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Approve purchase request",
        "tags": [
          "Purchase Requests"
        ]
      }
    },
...
    "/api/v1/spreadsheet/imports/latest": {
      "get": {
        "operationId": "SpreadsheetController_getLatestImport",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get latest spreadsheet import metadata",
        "tags": [
          "Spreadsheet"
        ]
      }
    },
    "/api/v1/spreadsheet/sheets": {
      "get": {
        "operationId": "SpreadsheetController_listSheets",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all imported sheets",
        "tags": [
          "Spreadsheet"
        ]
      }
    },
    "/api/v1/spreadsheet/sheets/{name}": {
      "get": {
        "operationId": "SpreadsheetController_getSheet",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get sheet metadata by name",
        "tags": [
          "Spreadsheet"
        ]
      }
    },
    "/api/v1/spreadsheet/sheets/{name}/rows": {
      "get": {
        "operationId": "SpreadsheetController_getSheetRows",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get sheet rows with all columns (paginated)",
        "tags": [
          "Spreadsheet"
        ]
      }
    },
...
  },
  "info": {
    "title": "ЦМК АВРОРА — ERP REST API",
    "description": "REST API v1 ЦМК АВРОРА — планирование производства, заказы, склад, финансы",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}